IN-381 Move assignments to a dedicated controller
+ add an endpoint to DELETE a single assignment
