+++ back/engines/commercial/insights/app/controllers/insights/web_api/v1/category_assignments_controller.rb
@@ -1,0 +1,70 @@
+# frozen_string_literal: true
+
+module Insights
+  module WebApi::V1
+    class CategoryAssignmentsController < ::ApplicationController
+      skip_after_action :verify_policy_scoped, only: :index # The view is authorized instead.
+
+      def index
+        render json: serialize_categories(input), status: :ok
+      end
+
+      def destroy
+        assignment = Insights::CategoryAssignment.find_by!(input: input, category_id: params.require(:id))
+        status = assignment.destroy.destroyed? ? :ok : 500
+        head status
+      end
+
+      # [PATCH] Adds new category assignments (idempotent).
+      def add_categories
+        assignments = assignment_service.add_assignments(input, categories)
+        errors = assignments.map(&:errors).select(&:any?)
+        if errors.present?
+          # TODO: Improvement: report all errors, not only the first one.
+          render json: { errors: errors.first.details }, status: :unprocessable_entity
+        else
+          render json: serialize_categories(input), status: :ok
+        end
+      end
+
+      def delete_categories
+        assignments = assignment_service.clear_all_assignments(input, view)
+        status = assignments.map(&:destroyed?).all? ? :ok : :internal_server_error
+        render status: status
+      end
+
+      private
+
+      def assignment_service
+        @assignment_service ||= Insights::CategoryAssignmentsService.new
+      end
+
+      # @return [Insights::View]
+      def view
+        @view ||= authorize(
+          View.includes(:scope).find(params.require(:view_id)),
+          :show?
+        )
+      end
+
+      def input
+        @input ||= view.scope.ideas.find(params.require(:input_id))
+      end
+
+      # @return [Array<Insights::CategoryAssignment>]
+      def categories
+        @categories ||= view.categories.find(
+          params.fetch(:data, []).select { |h| h[:type] == 'category' }.pluck(:id)
+        )
+      end
+
+      # @return [Hash]
+      def serialize_categories(input)
+        options = { include: %i[categories], params: fastjson_params }
+        InputSerializer.new(input, options)
+                       .serializable_hash
+                       .dig(:data, :relationships, :categories)
+      end
+    end
+  end
+end
