+++ back/engines/commercial/insights/app/controllers/insights/web_api/v1/inputs_controller.rb
@@ -12,35 +12,6 @@ module Insights
         render json: serialize(inputs)
       end
 
-      # [POST] Replaces category assignments altogether.
-      def replace_categories
-        ActiveRecord::Base.transaction do
-          assignment_service.clear_all_assignments(input, view)
-          assignment_service.add_assignments!(input, categories)
-        end
-        render json: serialize_categories(input), status: :ok
-      rescue ActiveRecord::RecordInvalid => e
-        render json: { errors: e.record.errors.details }, status: :unprocessable_entity
-      end
-
-      # [PATCH] Adds new category assignments (idempotent).
-      def add_categories
-        assignments = assignment_service.add_assignments(input, categories)
-        errors = assignments.map(&:errors).select(&:any?)
-        if errors.present?
-          # TODO: Improvement: report all errors, not only the first one.
-          render json: { errors: errors.first.details }, status: :unprocessable_entity
-        else
-          render json: serialize_categories(input), status: :ok
-        end
-      end
-
-      def delete_categories
-        assignments = assignment_service.clear_all_assignments(input, view)
-        status = assignments.map(&:destroyed?).all? ? :ok : :internal_server_error
-        render status: status
-      end
-
       private
 
       def assignment_service
@@ -63,13 +34,6 @@ module Insights
         @input ||= inputs.find(params.require(:id))
       end
 
-      # @return [Array<Insights::CategoryAssignment>]
-      def categories
-        @categories ||= view.categories.find(
-          params.fetch(:data, []).select { |h| h[:type] == 'category' }.pluck(:id)
-        )
-      end
-
       def serialize(inputs)
         options = {
           include: %i[categories suggested_categories source],
@@ -78,10 +42,6 @@ module Insights
         }
 
         InputSerializer.new(inputs, options)
-      end
-
-      def serialize_categories(input)
-        serialize(input).serializable_hash.dig(:data, :relationships, :categories)
       end
     end
   end
